--- vllm_kunlun/adapters/__init__.py
@@ -1,18 +1,0 @@
-"""Shared adapter infrastructure for optional, out-of-tree feature packs."""
-from .runtime_utils import (
-    WarningOnce,
-    env_bool,
-    env_int,
-    env_str,
-    find_op,
-    make_static_cpu_tensor,
-)
-
-__all__ = [
-    "WarningOnce",
-    "env_bool",
-    "env_int",
-    "env_str",
-    "find_op",
-    "make_static_cpu_tensor",
-]
