+++ vllm_kunlun/ops/attention/norms.py
@@ -23,7 +23,7 @@ from typing import Callable, List, Tuple
 import torch
 
 from ...ops.layernorm import KunlunRMSNorm
-from vllm_kunlun.adapters.runtime_utils import WarningOnce, find_op
+from vllm_kunlun.runtime_utils import WarningOnce, find_op
 
 LOGGER = logging.getLogger("vllm_kunlun.ops.attention.norms")
