+++ vllm_kunlun/ops/fused_moe/moe_hash_router.py
@@ -18,7 +18,7 @@ from typing import List, Tuple
 import torch
 import torch.nn.functional as F
 
-from vllm_kunlun.adapters.runtime_utils import WarningOnce
+from vllm_kunlun.runtime_utils import WarningOnce
 from vllm_kunlun.patches.registry import _register_lazy
 
 LOGGER = logging.getLogger("vllm_kunlun.ops.fused_moe.moe_hash_router")
